chore: Refactor <UserMenu /> to be reused on map and admin
