+++ packages/earth-admin/src/utils/hooks.tsx
@@ -18,32 +18,7 @@
 */
 
 import { add, compose } from 'lodash/fp';
-import { useEffect, useState } from 'react';
 import { useSWRInfinite } from 'swr';
-
-interface IError {
-  details: string;
-  message: string;
-}
-
-export function useDomWatcher(ref, callback, skip) {
-  useEffect(() => {
-    if (skip) {
-      return;
-    }
-
-    const handleClickOutside = (event) => {
-      if (ref.current && !ref.current.contains(event.target)) {
-        callback && callback();
-      }
-    };
-
-    document.addEventListener('mousedown', handleClickOutside);
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-    };
-  }, [ref, skip]);
-}
 
 /**
  * Custom hook that integrates useSWRInfinite with <DataListing /> component
