--- packages/earth-map/src/components/user-menu/styles.scss
@@ -1,69 +1,0 @@
-/*
-  Copyright 2018-2020 National Geographic Society
-
-  Use of this software does not constitute endorsement by National Geographic
-  Society (NGS). The NGS name and NGS logo may not be used for any purpose without
-  written permission from NGS.
-
-  Licensed under the Apache License, Version 2.0 (the "License"); you may not use
-  this file except in compliance with the License. You may obtain a copy of the
-  License at
-
-      https://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software distributed
-  under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
-  CONDITIONS OF ANY KIND, either express or implied. See the License for the
-  specific language governing permissions and limitations under the License.
-*/
-
-@import '~styles/config';
-
-.ng-user-account {
-  position: absolute;
-  top: 10px;
-  right: 20px;
-  z-index: 10000;
-}
-
-.ng-user-profile {
-  box-shadow: 0 0 5px 0 rgba($marapp-gray-8, 0.5);
-  border-radius: 50%;
-  height: 30px;
-  width: 30px;
-}
-
-.ng-user-profile-dropdown {
-  box-shadow: 0 0 5px 0 rgba($marapp-gray-100, 0.5);
-  background: $marapp-gray-7;
-  min-width: 175px;
-  padding: 25px !important;
-  position: absolute;
-  right: 0;
-  transform: translateY(15px);
-  li {
-    color: $marapp-gray-5;
-    font-family: $marapp-primary-font;
-    font-size: 13px;
-    font-weight: bold;
-    letter-spacing: 2.6px;
-    line-height: 25px;
-    padding-bottom: 10px;
-    &.ng-user-profile-signout, &.ng-user-profile-signin {
-      border-top: 1px solid $marapp-gray-8;
-      padding-top: 20px;
-      padding-bottom: 0;
-    }
-    a {
-      border-bottom: 0;
-      color: $marapp-gray-0;
-      font-family: $marapp-secondary-font;
-      font-size: 15px;
-      font-weight: 400;
-      letter-spacing: 0;
-      &:hover {
-        border-bottom: 1px solid $marapp-primary-color;
-      }
-    }
-  }
-}
