+++ packages/earth-shared/src/components/user-menu/index.ts
@@ -1,0 +1,1 @@
+export { UserMenu } from './UserMenu';
