+++ packages/earth-shared/src/components/user-menu/styles.scss
