+++ packages/earth-shared/src/index.js
@@ -21,3 +21,4 @@ export * from './components';
 export * from './authorization';
 export * from './services';
 export * from './utils';
+export * from './hooks';
